rename `api` package to `apiclient`
it's not an api, it's a client of an API!
