+++ fk/secretsend.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/fluidkeys/crypto/openpgp"
 	"github.com/fluidkeys/crypto/openpgp/armor"
-	"github.com/fluidkeys/fluidkeys/api"
+	"github.com/fluidkeys/fluidkeys/apiclient"
 	"github.com/fluidkeys/fluidkeys/colour"
 	"github.com/fluidkeys/fluidkeys/out"
 	"github.com/fluidkeys/fluidkeys/pgpkey"
@@ -40,7 +40,7 @@ import (
 func secretSend(recipientEmail string, filename string) exitCode {
 	armoredPublicKey, err := client.GetPublicKey(recipientEmail)
 	if err != nil {
-		if err == api.ErrPublicKeyNotFound {
+		if err == apiclient.ErrPublicKeyNotFound {
 			out.Print("\n")
 			out.Print("Couldn't find " + recipientEmail + " on Fluidkeys.\n\n")
 			out.Print("You can invite them to install Fluidkeys:\n")
